`git cl format` corrects #include
Having #include instead of #import in .mm files is now forbidden in
Chromium. It’s detected by presubmits. Given how easy it is to fix
automatically, I believe that, as most other trivial presubmit issues,
it should be corrected by `git cl format`.

Tested:
Manually modifying a file and running `git cl format --objective-c`.

It’s disabled by default, in order for volunteers to test the feature
up until it is deemed safed to turn it on by default.

Bug: 1352372
Change-Id: Iba2ce25b59a412b8083305b251e666cd45dad47d