Implement partial clones in gclient
While we have shallow clones in gclient it is not desirable in case the
git history is required. Starting from version 2.33 of git partial
clones are supported, that is, a git repository can be cloned by making
usage of filters to both speed up the process and consume less space.
But the downside of this is that git might download commits, objects
and trees on demand, though for large repositories with code search like
Chromium this might not affect things that much.

TEST=Fetch the Chromium sources using
`fetch --blobless --nohooks chromium` and synced changes with
`gclient sync --blobless`.

Change-Id: I3410205eef283f97f31bac507c5d64fb882a6da9