Document filter lists more precisely
The FilterSourceFile comment block says that the filter lists use
regular expressions, but that is not specific enough because .search and
.match have different semantics. crrev.com/c/3679163 added a filter that
only matched on PRESUBMIT.py instead of all of them - better
documentation might have avoided this.

The new documentation also makes it clear that Linux-style path
separators in the regex expressions are fine (they have been supported
since crrev.com/c/3609117).

Bug: 1309977
Change-Id: I49d37fbb93ef4a85f143d928a56af8ce22850c76