Avoid full fetches with gclient sync --no-history
This reworks the behavior of gclient sync --no-history to avoid fetching
the full repository history. This is accomplished by initializing a bare
repository and fetching the ref we are interested in.

The fetched depth is currently 5 to work around issues with commits
messages missing tags. This issue appears to only impact lastchange.py
which looks for Cr-Commit-Position.

Bug: 1305189
Change-Id: Ifea2a061308c354be15abd4e945141599ec20c9e