Speed up git thaw
Previously, git thaw would read the whole output of git rev-list HEAD
via readlines(). This was unnecessary, because we almost always only
need to look a few of the most recent commits. Most of the runtime of git-thaw was spend on this.

After this commit, we only read the lines we actually need. This makes git thaw run much faster.

Bug: 1378479
Change-Id: I6f6c06e1df55b4943a94642aa414fc11aeea5718