Replace gsutil.py with gcloud
This changes remove gsutil.py and replaces it with the updated gcloud command. Instead of downloading a pinned binary within the gsutil.py file, this implementation gets the version from cipd and has executable wrappers to invoke this.

This change also replaces every usage of gsutil.py with the new gcloud executable.

Bug: 959170, 1359383
Change-Id: Idcffb9250db61dd90d7e6d896d70d1f6fb962151