Handle goma_ctl/autoninja /?
With nested batch file it is very common for /? arguments to end up
being interpreted as a request for help on the batch file "call"
command. This is particularly problematic for commands like goma_ctl
and autoninja which don't print help if no arguments are passed, thus
making finding the correct help incantation slightly challenging.

This special-cases /? in these two batch files. Some other common batch
files were tested and found to be not affected, although this is
probably not a complete fix.

Change-Id: Ic9b0455c2f8b085666862bc4495d9bad445dfeaf