Update gsutil.py to use luci-auth for authentication
The change updates gsutil.py to internally use luci-auth for
authentication. It redirects `gsutil.py config` to `luci-auth login` and
wraps gsutil.py calls using `luci-auth context`.

The change is mainly done to ensure users have a way to authenticate once OOB
authentication flow is deprecated. This patch falls back to using the .boto
file when that is provided.

Bug: 959170, 1359383
Change-Id: Icc6df1d8b93424ac0f6e2df84afdebc869763ede