Account for managed solution in src when finding ninja
At present when looking for ninja, main only searches in two locations, gclient root and src. In embedders projects, like Electron, it's the case that the managed solution would live inside src, and so we also need to look inside src to potentially find ninja. This is currently causing issues in Electron.

Change-Id: I35aea29135d5968722a308bfb6dbb8a727550b42