Reset git_cache mirror when origin/HEAD is missing
Due to master->main branch migration, some of the builders have their
git cache in an invalid state, i.e., missing origin/HEAD ref. This leads
to bot_update failure in several builders. To overcome this, this change
adds a check to git_cache to verify if this ref exist, and if not, drop
the mirror.

This change has been tested on a copy of failing infra_internal cache
from one of the bots.

Bug: 1418866
Change-Id: Id716a013833d459ee37306d27b369099faf846ab