Disable two encoding tests on Windows
test_unicode and test_default_encoding fail on Windows because the
default encoding is (typically) cp1252. I don't think this is
generically fixable at this point (encodings must be specified when
opening files) so these tests should be disabled on Windows.

This this change "git cl presubmit --all" runs without errors on
depot_tools on Windows.

Change-Id: I043e09e82c172b93761dc06c77784cdc7d3e48aa