Reset alternates when the git_cache dir is updated
When cache URL is updated, `gclient sync` appends the new cache path as
an additional alternates intead of overriding it. This was done to
support change in mirror path. However, when the cache is reset, gclient
appends the new mirror path instead of overwriting the existing one.
This results in failed builds where `git fetch` looks for non-existent
alternate object dbs.

Bug: 1428312
Change-Id: Ib10e0b966ddef8793415cf8ef0eddbec13ba1a8d