Suppress stderr of the git rev-parse command
In GetPrimarySolutionPath(), redirect the stderr of the
git rev-parse --show-toplevel command to /dev/null. This avoids the
following error message when running depot_tools' ninja command outside
the Chrome source tree:
fatal: not a git repository (or any of the parent directories): .git

I believe this error message was an unintended side effect of the work
on bug 1340825.

Bug: 1340825
Change-Id: I5679a854d1ff23f581564c54c765e13f8ba8b534