Add convenience wrapper for reclientreport for autoninja users
This is designed to be called by a developer when they want to submit an reclient bug report.
Developers could just call the //buildtools/reclient/reclientreport binary directly but this wrapper provides an easier way to set the same log flags that are set by ninja_reclient.
I dont think we should bundle this logic into autoninja as this is not required in 99% of builds.

Bug: b/277763387
Change-Id: I0204b6acc22f199b461ef710d0dfd07e05534af7