Warn on Fixes:, and suggest Fixed:
The Bug: syntax is well known as a way of tagging bugs in a commit
message, at least partially because when it is missing there is a
presubmit warning suggesting that it be used.

The Fixed: syntax is less well known because there is no such reminder.
And, if somebody types "Fixes: " instead of "Fixed: " there is no error.
Therefore it is not surprising that "Fixes: " was used seven times in
the last 10,000 commits. See crrev.com/c/4416218 for one example.

This change adds both a reminder that Fixed: exists as well as an error
if Fixes: is used.

Fixed: 1434702
Change-Id: Id940d0af5cfea206ee923fbb779bf686b34c5f33