Allow 409 error from Gerrit when creating the branch
Gerrit may have a longer responding time. If it hits our timeout,
gerrit_client will retry but may get 409 error, which means the
branch turns out was created.

We should ignore this error if the branch head is the commit we
requested. Also pre-check the branch status before creating it.
If an existing branch is expected by the caller, simply return
the branch info(if already created) without sending the creation
request.

BUG=1433047
TEST=train
Recipe-Nontrivial-Roll: chrome_release
Change-Id: Ic6184705579f1410cb9c54843cc0bbb4c4ad61b1