Allow calling tag functions from PostUploadHook
Calling CheckChangeHasNoUnwantedTags and CheckChangeHasBugFieldFromChange
from a PostUploadHook is tempting because it ensures that users will see
a message if they add a bad tag when creating the commit message during
upload. Allowing this requires creating entry points for these functions
that take a change parameter instead of an input_api parameter. It also
requires fixing our use of the deprecated inspect.getargspec function.

Note that the getargspec change is only easily done now that Python 2
is no longer supported.

This change is required in support of crrev.com/c/4490087.

Bug: 1434702
Change-Id: I9517f989e17f9f18305680c60e0c7a020f7c7519