Allow 409 error from Gerrit when creating the branch
Gerrit may return 409 error to us even if the branch was created.
We should ignore this error if the branch head is the commit we
requested.

Also pre-check the branch status before creating it.
If an existing branch is expected by the caller, simply return
the branch info without sending the creation request.

BUG=1433047
TEST=local run
Change-Id: Ic6184705579f1410cb9c54843cc0bbb4c4ad61b1