unset %PSModulePath% before calling powershell
pwsh, cmd and powershell interoperability breaks `Get-FileHash` cmdlet
which is used in .cipd_impl.ps1. The issue is a result of modified
%PSModulePath% by pwsh which is incorrectly being consumed by powershell
later causing the cmdlet to be not found.

This CL tries to workaround the issue by unsetting %PSModulePath% before
calling powershell which forces it to use its default %PSModulePath%.

Bug: 1443163
Change-Id: I117e5077239c5371bb392b36b521671095029abf