unset %PSModulePath% before calling powershell
pwsh, cmd and powershell interoperability breaks `Get-FileHash` cmdlet
which is used in .cipd_impl.ps1. The issue prevents loading of correct
PowerShell module due to how $PSModulePath is structured in some cases.

With this CL, we always import the compatible module at the beginning
of the script. This ensures Get-FileHash is always available to us
regardless of the PowerShell version being used.

Bug: 1443163
Change-Id: I117e5077239c5371bb392b36b521671095029abf