Update TODO pattern to not match other names
We plan on introducing a TODOC macro in v8.
(This is similar to a TODO but as a note to document something.)
At the current state the linter complains on lines like:
// TODOC: [...]
claiming that the TODO misses a username.

We'd like our new macro not to require a username.
The new regex is a merge of the existing regex with the updated
regex in the google-internal source this script is based on.

Change-Id: Ib792c82a9d62e45ef7994beab3fb3f48d7241292