[gclient] Export CONFIG_CTX for other modules to extend.
The recipe engine has a refactor coming up where CONFIG_CTX needs to
be explicitly exported for other modules to extend it. Currently
a handful of modules in the build, build_internal and
release_scripts repos extend the gclient configuration context.

Exporting it like this should have no functional effect, but is just
in preparation for the upcoming refactor.

R=gavinmak

Bug: 1462728
Change-Id: I5871d472f7e28d38eefa9b1e89cfcc804823a1bd