[git-cl] Greatly improve the performance of `git config` operations.
I observed that `git map-branches` was particularly slow in a repo
I have which has ~40 local branches or so. Previous to this change
I saw:

    $ time git map-branches -vvv
    ...
    git map-branches -vvv  4.14s user 4.60s system 45% cpu 19.280 total
    $ time git map-branches -vv
    ...
    git map-branches -vv  2.13s user 2.17s system 41% cpu 10.336 total

And after this change I see:

    $ time git map-branches -vvv
    ...
    git map-branches -vvv  1.17s user 0.94s system 47% cpu 4.478 total
    $ time git map-branches -vv
    ...
    git map-branches -vv  0.43s user 0.16s system 42% cpu 1.394 total

Which is a speedup of between ~4-7x, depending on how much verbosity
you want. As a bonus, `-v` and `-vv` now take the exact same amount
of time to run (so users can upgrade from `-v` to `-vv` without
penalty), and non-verbose-mode in this checkout takes a mere 1.1s on
average.

I suspect this change will also speed up all general `git-cl`
operations which do many interrogations of the git config,
especially on Windows, where process execution overhead is
relatively severe.

Change-Id: I0af8b5fe29612212ff326b6c83e8cc1cab25f63a