gclient: Call _AutoFetchRef before using a rev in _FetchAndReset
Other places in gclient_scm (GitWrapper._Clone, GitWrapper.update) do the same.

Note that we call _AutoFetchRef even after running _Fetch, because _Fetch in not
guaranteed to fetch a specific revision. For example, a checkout might be
configured without branch heads and without tags, yet a revision might point to
a tag.

Change-Id: I5cfd8ebafa4b45afd72b0251f04e5d713199f2fd