Teach autoninja to handle siso
autoninja is intended to mean "build this directory, don't bother me
with the details". As we transition to siso it is therefore appropriate
that it should take on the duty of deciding whether to invoke ninja or
siso.

By looking for a use_siso gn arg autoninja can easily do this.

To make this work a dummy use_siso gn arg needs to be added to one of
the gn files in Chromium's build directory.

This change also teaches autoninja to detect if a user switches
between siso and ninja without doing a gn clean inbetween.

Note that this change teaches autoninja to invoke siso when appropriate,
but this doesn't actually work - siso can't handle the -C option needed
for builds.

Bug: b/293657720
Change-Id: I3ad36a81857e75ffe6babc4f107b777e733a285b