Teach autoninja to handle siso/autosiso
autoninja is intended to mean "build this directory, don't bother me
with the details". As we transition to siso it is therefore appropriate
that it should take on the duty of deciding whether to invoke ninja or
siso.

By looking for a use_siso gn arg autoninja can easily do this.

This change relies on crrev.com/c/4753433 to add the use_siso gn arg.

This change also teaches autoninja to detect if a user switches
between siso and ninja without doing a gn clean inbetween.

Note that this change also teaches autoninja to invoke autosiso or
siso ninja based on whether use_remoteexec is true.

Bug: b/293657720
Change-Id: I3ad36a81857e75ffe6babc4f107b777e733a285b